Python Task/web/requirements.txt
pika
streamlit
pandas
streamlit-autorefresh
psycopg2-binary 